+++ pulsar/apps/rpc/__init__.py
@@ -5,3 +5,4 @@ from .exceptions import *
 from .handlers import *
 from .jsonrpc import *
 from .decorators import *
+from .mixins import *