+++ pulsar/apps/rpc/mixins.py
@@ -1,9 +1,13 @@
-from pulsar.http import rpc
 from pulsar.utils.py2py3 import iteritems
 
+from .jsonrpc import JSONRPC
 
-class PulsarServerCommands(rpc.JSONRPC):
-    '''Some useful commands to use as plugin into your rpc server'''
+
+__all__ = ['PulsarServerCommands']
+
+
+class PulsarServerCommands(JSONRPC):
+    '''Some useful commands to get you started.'''
     
     def rpc_ping(self, request):
         '''Ping the server'''
@@ -19,6 +23,12 @@ class PulsarServerCommands(rpc.JSONRPC):
     def rpc_functions_list(self, request):
         return list(self.listFunctions())
     
+    def rpc_kill_actor(self, request, aid):
+        '''Kill and actor which match the id *aid*'''
+        # get the worker serving the request
+        worker = request.environ['pulsar.worker']
+        worker.proxy.kill_actort(worker.arbiter, aid)
+        
     def extra_server_info(self, request, info):
         return info